eliminate mypy errors

+++ speech/project3/hmm_test.py
@@ -4,7 +4,7 @@ import unittest
 
 import numpy as np
 
-from speech.project3.hmm import HMM, HMM_Single
+from speech.project3.hmm import HMM_Single
 
 
 class TestAudio(unittest.TestCase):
@@ -13,12 +13,14 @@ class TestAudio(unittest.TestCase):
         covariance = np.eye(39)
         num_samples = 200
         np.random.seed(0)
-        samples = np.random.multivariate_normal(mean, covariance, num_samples)
+        samples = np.random.multivariate_normal(mean, covariance, num_samples).astype(
+            np.float32
+        )
         s = [samples[i : i + 40] for i in range(0, 161, 10)]
         dtw = HMM_Single(s, 5, 1)
         another_samples = np.random.multivariate_normal(
             np.random.normal(2, 1, 39), np.eye(39) * 1, 200
-        )
+        ).astype(np.float32)
         a_s = [another_samples[i : i + 40] for i in range(0, 161, 10)]
         for i in range(10):
             print(dtw.predict_score(s[i]))
